Third wave of major changes. Binary support now fully implemented (at least covering every contingency I can think of). Utility functions extracted to their own library.
