+++ meerwatch_tools.py
@@ -18,8 +18,9 @@ import subprocess as sproc
 import matplotlib.pyplot as plt
 from matplotlib import colors, cm
 from astropy.time import Time
-from decimal import Decimal,InvalidOperation
-from scipy.optimize import fsolve
+
+# Custom imports
+from binary_tools import is_binary, get_binphase
 
 # Constants
 DAYPERYEAR = 365.25
@@ -285,211 +286,3 @@ def plot_toas_fromarr(toas, pid="unk", mjd=None, fs=14, out_file="toas.png", out
     plt.savefig(out_file, bbox_inches='tight')
 
     plt.clf()
-
-# --- BINARY PHASE CODE ---
-# --- Largely adapted from Scintools code provided by Daniel Reardon ---
-
-# reads a par file into a dictionary object
-# this functionality is already performed to an extent by PSRDB / util.ephemeris, but
-# that code is moreso just to read and store the fields. This code has better control for
-# handling/mainpulating/calculating with those fields, and so for the moment I'll retain it.
-def read_par(parfile):
-    """
-    Reads a par file and return a dictionary of parameter names and values
-    """
-
-    par = {}
-    ignore = ['DMMODEL', 'DMOFF', "DM_", "CM_", 'CONSTRAIN', 'JUMP', 'NITS',
-              'NTOA', 'CORRECT_TROPOSPHERE', 'PLANET_SHAPIRO', 'DILATEFREQ',
-              'TIMEEPH', 'MODE', 'TZRMJD', 'TZRSITE', 'TZRFRQ', 'EPHVER',
-              'T2CMETHOD']
-
-    file = open(parfile, 'r')
-    for line in file.readlines():
-        err = None
-        p_type = None
-        sline = line.split()
-        if len(sline) == 0 or line[0] == "#" or line[0:2] == "C " or sline[0] in ignore:
-            continue
-
-        param = sline[0]
-        if param == "E":
-            param = "ECC"
-
-        val = sline[1]
-        if len(sline) == 3 and sline[2] not in ['0', '1']:
-            err = sline[2].replace('D', 'E')
-        elif len(sline) == 4:
-            err = sline[3].replace('D', 'E')
-
-        try:
-            val = int(val)
-            p_type = 'd'
-        except ValueError:
-            try:
-                val = float(Decimal(val.replace('D', 'E')))
-                if 'e' in sline[1] or 'E' in sline[1].replace('D', 'E'):
-                    p_type = 'e'
-                else:
-                    p_type = 'f'
-            except InvalidOperation:
-                p_type = 's'
-
-        par[param] = val
-        if err:
-            par[param+"_ERR"] = float(err)
-            
-        if p_type:
-            par[param+"_TYPE"] = p_type
-
-    file.close()
-
-    return par
-
-def get_binphase(mjds, pars):
-    """
-    Calculates binary phase for an array of barycentric MJDs and a parameter dictionary
-    """
-    
-    U = get_true_anomaly(mjds, pars)
-
-    OM = get_omega(pars, U)
-
-    # normalise U
-    U =np.fmod(U, 2*np.pi)
-
-    return np.fmod(U + OM + 2*np.pi, 2*np.pi)/(2*np.pi)
-
-def get_ELL1_arctan(EPS1, EPS2):
-    """
-    Given the EPS1 and EPS2 parameters of the ELL1 binary model,
-    calculate the arctan(EPS1/EPS2) value accounting for all degeneracies and ambiguities.
-    This function has been abstracted as it is needed for calculating both OM and T0
-    """
-
-    # check for undefined tan
-    if (EPS2 == 0):
-        if (EPS1 > 0):
-            AT = np.pi/2
-        elif (EPS1 < 0):
-            AT = -np.pi/2
-        else:
-            # eccentricity must be perfectly zero - omega is therefore undefined
-            AT = 0
-    else:
-        AT = np.arctan(EPS1/EPS2)
-        # check for tan degeneracy
-        if (EPS2 < 0):
-            AT += np.pi
-
-    return np.fmod(AT + 2*np.pi, 2*np.pi)
-
-def get_omega(pars, U):
-    """
-    Calculate the instantaneous version of omega (radians) accounting for OMDOT
-    per Eq. 8.19 of the Handbook. May be slightly incorrect for relativistic systems
-    """
-
-    # get reference omega
-    if 'TASC' in pars.keys():
-        if 'EPS1' in pars.keys() and 'EPS2' in pars.keys():
-
-            OM = get_ELL1_arctan(pars['EPS1'], pars['EPS2'])
-            # ensure OM within range 0..2pi
-            OM = np.fmod(OM + 2*np.pi, 2*np.pi)
-
-        else:
-            OM = 0
-    else:
-        if 'OM' in pars.keys():
-            OM = pars['OM'] * np.pi/180
-        else:
-            OM = 0
-
-    # get change in omega
-    if 'OMDOT' in pars.keys():
-        # convert from deg/yr to rad/day
-        OMDOT = pars['OMDOT'] * (np.pi/180) / DAYPERYEAR
-    else:
-        OMDOT = 0
-    
-    # calculate current omega
-    PB = pars['PB'] # days
-
-    OM = OM + OMDOT*U/(2*np.pi/PB)
-    
-    return OM
-
-def get_true_anomaly(mjds, pars):
-    """
-    Calculates true anomalies for an array of barycentric MJDs and a parameter dictionary
-    """
-
-    # handle orbital period
-    PB = pars['PB']  # days
-
-    if 'PBDOT' in pars.keys():
-        PBDOT = pars['PBDOT']
-    else:
-        PBDOT = 0
-
-    if np.abs(PBDOT) > 1e-6: # adjusted from Daniels' setting
-        # correct tempo-format
-        PBDOT *= 10**-12
-
-    NB = 2*np.pi/PB
-
-    # handle ELL1 ephemeris
-    if 'TASC' in pars.keys():
-        if 'EPS1' in pars.keys() and 'EPS2' in pars.keys():
-            T0 = pars['TASC'] + get_ELL1_arctan(pars['EPS1'], pars['EPS2'])/NB  # MJD - No PBDOT correction required as referenced to zero epoch
-            ECC = np.sqrt(pars['EPS1']**2 + pars['EPS2']**2)
-        else:
-            T0 = pars['TASC']
-            ECC = 0
-    else:
-        T0 = pars['T0']  # MJD
-        if 'ECC' in pars.keys():
-            ECC = pars['ECC']
-        else:
-            ECC = 0
-
-    # mean anomaly
-    M = NB*((mjds - T0) - 0.5*(PBDOT/PB) * (mjds - T0)**2)
-    M = M.squeeze()
-
-    # eccentric anomaly
-    if ECC < 1e-4:
-        print('Assuming circular orbit for true anomaly calculation')
-        E = M
-    else:
-        M = np.asarray(M, dtype=np.float64)
-        E = fsolve(lambda E: E - ECC*np.sin(E) - M, M)
-        E = np.asarray(E, dtype=np.float128)
-
-    # true anomaly
-    U = 2*np.arctan2(np.sqrt(1 + ECC) * np.sin(E/2), np.sqrt(1 - ECC) * np.cos(E/2))  # true anomaly
-
-    if hasattr(U,  "__len__"):
-        U[np.argwhere(U < 0)] = U[np.argwhere(U < 0)] + 2*np.pi
-        U = U.squeeze()
-    elif U < 0:
-        U += 2*np.pi
-
-    # final change - need to have U count the number of orbits - rescale to match M and E
-    E_fac = np.floor_divide(E, 2*np.pi)
-    U += E_fac * 2*np.pi
-
-    return U
-
-def is_binary(pars):
-    """
-    Determine if a set of parameters adequately describes a binary pulsar
-    """
-
-    retval = False
-
-    if ('BINARY' in pars.keys() and 'PB' in pars.keys() and ('TASC' in pars.keys() or 'T0' in pars.keys())):
-        retval = True
-
-    return retval
